usePromise detects unmounted components.
Currently, if a component using usePromise (or useFetch, which uses usePromise) is unmounted before the encapsulated work is complete. It will still try to update the react state and trigger an annoying warning in react dev mode.

This wraps the state updates inside a check to see if the component is still mounted.